+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - `unnecessary_indexing_checker` has now been extended to check comprehensions in addition to for loops.
 - `invalid_for_target_checker` has now been extended to check comprehensions in addition to for loops.
-- Violated representation invariant error message now includes the current values of the instance attributes.
+- Violated representation invariant error message now includes the class name and current values of the instance attributes.
 
 ### Bug Fixes